+++ Tests/DependencyInjection/Fixtures/config/php/many_mailers.php
@@ -14,14 +14,17 @@ $container->loadFromExtension('pixelart_swiftmailer_manipulator', [
         'first_mailer' => [
             'prepend_subject' => '[TESTSYSTEM 1!]',
             'prepend_body' => 'swiftmailer/prepend_body_1.txt.twig',
+            'from_address' => 'fake_1@example.com',
         ],
         'secondary_mailer' => [
             'prepend_subject' => '[TESTSYSTEM 2!]',
             'prepend_body' => 'swiftmailer/prepend_body_2.txt.twig',
+            'from_address' => 'fake_2@example.com',
         ],
         'third_mailer' => [
             'prepend_subject' => '[TESTSYSTEM 3!]',
             'prepend_body' => 'swiftmailer/prepend_body_3.txt.twig',
+            'from_address' => 'fake_3@example.com',
         ],
     ],
 ]);