+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Impersonate plugin to modify the from address of a message from @angelov
+
 ### Internal changes
 - Refactor the service name to support multiple plugins